create, read and edit are working

+++ backend/src/index.ts
@@ -3,14 +3,19 @@ import express, { Router } from "express";
 import cookieParser from "cookie-parser";
 import db from "./db";
 import jwt, { decode, sign, verify } from "jsonwebtoken"
-import { User } from "./models";
+import { Todo, User } from "./models";
 import { compare, genSalt, hash } from "bcrypt"
 import authMiddleware from "./middleware";
-
+import cors from "cors";
 const app = express();
 
 app.use(express.json())
 app.use(cookieParser())
+app.use(cors({
+    methods: ["POST", "GET", "OPTIONs", "DELETE"],
+    credentials: true,
+    origin: "http://localhost:5500"
+}))
 
 app.post("/login", async (req, res) => {
 
@@ -33,7 +38,12 @@ app.post("/login", async (req, res) => {
         });
     }
     const token = sign({ id: user.id, email: user.email }, process.env.JWT_SECRET || "HELLO");
-    return res.cookie("token", token).status(200).json({ "message": "Logged in successfully" });
+    if(req.cookies){
+        req.cookies["token"] = ""
+    }
+    return res.cookie("token", token, {
+        httpOnly: true, secure: false, sameSite: 'lax', maxAge: 1000 * 60 * 60 * 24 * 7 // 7 days
+    }).status(200).json({ "message": "Logged in successfully" });
 })
 
 app.post("/signup", async (req, res) => {
@@ -74,10 +84,11 @@ protectedRoutes.post("/create", async (req, res) => {
     }
     const id = (decode(req.cookies.token) as { id: number, email: string }).id
     try {
-        const result = await db.query("INSERT INTO todos (title, description, user_id) VALUES ($1,$2,$3)", [title, description, id]);
+        const result = await db.query<Todo>("INSERT INTO todos (title, description, user_id) VALUES ($1,$2,$3) RETURNING *", [title, description, id]);
         if (result.rowCount && result.rowCount === 1) {
             res.status(201).json({
-                "message": "Todo Created Successfully"
+                "message": "Todo Created Successfully",
+                "todo":result.rows[0]
             })
         }
     } catch (e) {
@@ -88,8 +99,12 @@ protectedRoutes.post("/create", async (req, res) => {
 
 protectedRoutes.delete("/delete", async (req, res) => {
     const { id } = req.body
+    if (!id) {
+        return res.status(400).json({
+            "error": "Invalid data"
+        })
+    }
     const user_id = (decode(req.cookies.token) as { id: number, email: string }).id
-
     try {
         const result = await db.query("DELETE FROM todos WHERE id = $1 AND user_id = $2", [id, user_id]);
         if (result.rowCount && result.rowCount === 1) {
@@ -103,17 +118,22 @@ protectedRoutes.delete("/delete", async (req, res) => {
     }
 })
 
-protectedRoutes.post("/edit",async (req, res) => {
+protectedRoutes.post("/edit", async (req, res) => {
     const { id, title, description } = req.body
+    if (!id || !title || !description) {
+        return res.status(400).json({
+            "error": "Invalid data"
+        })
+    }
     const user_id = (decode(req.cookies.token) as { id: number, email: string }).id
     try {
         const query = "UPDATE todos SET title = $1, description = $2,updated_at = NOW() WHERE id = $3 AND user_id = $4 RETURNING *"
         const result = await db.query(query, [title, description, id, user_id]);
-        if(result.rowCount && result.rowCount === 0){
+        if (result.rowCount && result.rowCount === 0) {
             return res.status(404).json({ error: "Todo not found for user" });
         }
         return res.status(200).json({
-            "message":"todo updated successfully"
+            "message": "todo updated successfully"
         })
     } catch (e) {
         console.log(e)
@@ -121,20 +141,20 @@ protectedRoutes.post("/edit",async (req, res) => {
     }
 })
 
-protectedRoutes.get("/all", async(req,res) => {
+protectedRoutes.get("/all", async (req, res) => {
     const user_id = (decode(req.cookies.token) as { id: number, email: string }).id
-    try{
-        const result = await db.query("SELECT * FROM todos WHERE user_id = $1",[user_id]);
-        if(result.rowCount && result.rowCount > 0){
+    try {
+        const result = await db.query("SELECT * FROM todos WHERE user_id = $1", [user_id]);
+        if (result.rowCount && result.rowCount > 0) {
             return res.status(200).json({
-                "todos":result.rows
+                "todos": result.rows
             })
         } else {
             return res.status(200).json({
-                "message":"no todos were found"
+                "message": "no todos were found"
             })
         }
-    } catch(e){
+    } catch (e) {
         return res.status(500)
     }
 })
